+++ package/adapters/__init__.py
@@ -1,0 +1,6 @@
+from .LanguageAstAdapter import LanguageAstAdapter
+from .LanguageAstAdapter import NodeType
+from .PythonAstAdapter import PythonAdapter
+from .CppAstAdapter import CppAstAdapter
+from .ErlangAstAdapter import ErlangAdapter
+from .LanguageAstAdapterRegistry import LanguageAstAdapterRegistry
